requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.37.22
botocore==1.37.22
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.0
gradio_client==1.8.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
jmespath==1.0.1
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tqdm==4.67.1
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
